provides default ppn for batch systems
This may not be required, but it is useful to developers who explicitly
require an oxymoron batch. The ppn for an Oxymoron batch job is unique
and can be confusing. This hopes to alleviate it.

There is a possibility this should be moved to osc-vnc as I want to keep
this gem as low level as possible. Oh well :(

This fixes #2