don't break up job arrays
A job array initially starts off as a single job with a single job id.
But as one of those jobs in the array changes state (e.g., runs) then it
will be given a new job id and be treated as a separate job. This new
job will be on a new line in `squeue` and we don't need to use
`--array`.