Addition of DRMAA assisted option for getting info on a single job
There are now two flows for SGE#info. The original where qstat is called without a job id filter is used if the user does not provide :sge_root. The alternative flow depends on DRMAA being installed. If it is then qstat is called with a job filter, and the job's missing status is filled in using DRMAA. There is an optional configuration option to set LD_LIBRARY_PATH if necessary.

Removed additional qacct related code