Adds an extension to the cluster configs to allow scheduler client overrides
Quoting #111:

It seems we have several use cases where being able to provide a custom wrapper around qsub or sbatch, without having to wrap the rest of the commands, would be useful. Currently you can only specify path to bin in the cluster config. The idea is you [can now] do this:

```yaml
  job:
    adapter: "torque"
    host: "owens-batch.ten.osc.edu"
    lib: "/opt/torque/lib64"
    bin: "/opt/torque/bin"
    custom_bin:
      qsub: "/opt/torque-osc-custom/bin/qsub"
    version: "6.0.1"
```