Fix SGE adapter crash: DRMAA constant not defined
A test was using the too general "StandardError" to
verify that OodCore::Job::Adapters::Sge::Batch::Error was raised.

As a result, when NameError: uninitialized constant OodCore::Job::Adapters::Sge::Batch::DRMAA
was being raised the test would not catch this. This occurs not using libdrmma.

We change the test to verify the specific Batch::Error exception is raised.
Then we fix the code to handle the missing DRMAA constant resulting in a crash.