Update LSF to handle job arrays
- Updated tests

One of the major quirks with LSF is that job array requests are appended to the job name. This is unusual and complicates creating a proxy parent because the array spec would have to be reconstructed from child tasks. It is conceviable that creating the original array spec might not be possible because child jobs might no longer be reported by LSF. Also, new since LSF version 9 bjobs supports a format string. This could be leveraged to simplify record parsing in future versions of the adapter.

The other quirk of LSF is that (at least in the workbench environment) the cluster requires a select statement to specify on what cluster resources will the jobs run. A switch has been added to the live cluster test to permit adding scheduler-specific native options.