Adds the start of a Singularity based adapter
Singularity's concept of services are interesting in that they can be rediscovered after being started using the command `singularity instance list --json`. What is frustrating about this is that there is no ability for an image started as a service to finish.

Take for example the Gateways team's documentation builder image (which uses Docker). This image starts up, performs a task, and exits. Singularity can do this using `singularity run` but that command blocks which does not map well to the batch work paradigm for which the Job Adapters were designed. In addition to blocking Singularity's run were it nohup'd and backgrounded offers no ability to "tag" the process for re-discovery.