Update Script#copy_environment to default nil to permit existing behavior to continue
- Add back missing Slurm environment setting
- Implement environment persistence for LSF, PBSPro, Torque and Grid Engine

I also noticed that multiple adapters when dealing with the contents of Script#job_environment explicitly stringify both the keys and the values; this is redundent as that mutation is performed in the Script constructor. Job submission always requires a Script object, which are immutable, and so the adapter-specific stringification sections in the code are safe to remove.