rm oidc token
We cannot use odic token with the --token flag. We'll have to
generate the configuration in the nginx stage due to timing issues
(the access_token is only valid for so long) and oidc realm issues
(the access_token given is for ood not the k8s cluster, it's a
different realm).