Kubernetes Adapter (#156)
Add Kubernetes adapter.

Add a kubernetes adapter to support interactive applications (not batch jobs).  This adapter supports GKE and OIDC auth types where it does some initialization to the ~/.kube/config to set a context to use. It can also do nothing and just use the kubeconfig as is. 

It currently submits a pod with accompanying service, configmap, and secret.  The configmap configures the pod. The service allows network access to pod. And the secret is the password configuration for the application running in the pod.  

The secret password and host are passed back through native attributes which differs from other adapters and is subject to change in future versions.

It also relies heavily on `script.native` attributes to determine what container to schedule (and how) along with any init containers as well and this too is subject to change.